# Remove derivative dependency