--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -21,7 +21,8 @@
 - On X11, prevent stealing input focus when creating a new window.
   Only steal input focus when entering fullscreen mode.
 - On Wayland, add support for set_cursor_visible and set_cursor_grab.
-- On Wayland, fixed DeviceEvents for relative mouse movement is not always produced
+- On Wayland, fixed DeviceEvents for relative mouse movement is not always produced.
+- Removed `derivative` crate dependency.
 
 # 0.20.0 Alpha 3 (2019-08-14)