--- a/Cargo.toml
+++ b/Cargo.toml
@@ -19,7 +19,6 @@ lazy_static = "1"
 libc = "0.2"
 log = "0.4"
 serde = { version = "1", optional = true, features = ["serde_derive"] }
-derivative = "1.0.2"
 raw-window-handle = "0.2"
 
 [dev-dependencies]