--- a/src/lib.rs
+++ b/src/lib.rs
@@ -121,8 +121,6 @@ extern crate log;
 #[macro_use]
 extern crate serde;
 #[macro_use]
-extern crate derivative;
-#[macro_use]
 #[cfg(any(target_os = "ios", target_os = "windows"))]
 extern crate bitflags;
 #[cfg(any(target_os = "macos", target_os = "ios"))]